+++ spikee/plugins/prompt_decomposition.py
@@ -17,7 +17,13 @@ import random
 from dotenv import load_dotenv
 
 from spikee.templates.plugin import Plugin
-from spikee.utilities.llm import get_example_llm_models, get_supported_llm_models, get_supported_prefixes, get_llm, validate_llm_option
+from spikee.utilities.llm import (
+    get_example_llm_models,
+    get_supported_llm_models,
+    get_supported_prefixes,
+    get_llm,
+    validate_llm_option,
+)
 
 
 class PromptDecompositionPlugin(Plugin):
@@ -26,7 +32,11 @@ class PromptDecompositionPlugin(Plugin):
     DEFAULT_MODE = "dumb"
 
     # Supported modes
-    SUPPORTED_MODES = [DEFAULT_MODE] + [model for model in get_example_llm_models()] + [model for model in get_supported_llm_models() if model != "offline"]
+    SUPPORTED_MODES = (
+        [DEFAULT_MODE]
+        + [model for model in get_example_llm_models()]
+        + [model for model in get_supported_llm_models() if model != "offline"]
+    )
 
     def get_available_option_values(self) -> List[str]:
         """Return supported options; first option is default."""